+++ Apizr/Src/Apizr/Configuring/Manager/ApizrManagerOptionsBuilder.cs
@@ -141,7 +141,7 @@ namespace Apizr.Configuring.Manager
                         {
                             var cacheSection = config.GetChildren().ToList();
                             var modeValue = cacheSection.FirstOrDefault(c => c.Key == "Mode")?.Value;
-                            var mode = !string.IsNullOrEmpty(modeValue) ? (CacheMode) Enum.Parse(typeof(CacheMode), modeValue) : CacheMode.GetAndFetch;
+                            var mode = !string.IsNullOrEmpty(modeValue) ? (CacheMode) Enum.Parse(typeof(CacheMode), modeValue) : CacheMode.FetchOrGet;
                             var lifeSpanValue = cacheSection.FirstOrDefault(c => c.Key == "LifeSpan")?.Value;
                             var lifeSpan = !string.IsNullOrEmpty(lifeSpanValue) ? TimeSpan.Parse(lifeSpanValue) : TimeSpan.Zero;
                             var shouldInvalidateOnErrorValue = cacheSection.FirstOrDefault(c => c.Key == "ShouldInvalidateOnError")?.Value;
@@ -806,7 +806,7 @@ namespace Apizr.Configuring.Manager
         }
 
         /// <inheritdoc />
-        public IApizrManagerOptionsBuilder WithCaching(CacheMode mode = CacheMode.GetAndFetch, TimeSpan? lifeSpan = null,
+        public IApizrManagerOptionsBuilder WithCaching(CacheMode mode = CacheMode.FetchOrGet, TimeSpan? lifeSpan = null,
             bool shouldInvalidateOnError = false)
         {
             Options.CacheOptions[ApizrConfigurationSource.ManagerOption] = new CacheAttribute(mode, lifeSpan, shouldInvalidateOnError);